Separate model into one dir.
