+++ .gitignore
@@ -1,2 +1,3 @@
 # Created by .gitignore support plugin (hsz.mobi)
 .idea
+*.pyc
