+++ user_map/models/__init__.py
@@ -1,0 +1,3 @@
+from user_map.models.user import *
+from user_map.models.role import *
+from user_map.models.user_manager import *
